use context for config
