+++ src/components/map/mainMap/index.jsx
@@ -1,8 +1,8 @@
 import { MapboxProvider } from '../../../context/mapContext';
 
-export const MainMap = ({ children,config }) => {
+export const MainMap = ({ children }) => {
   return (
-    <MapboxProvider config={config}>
+    <MapboxProvider>
       {/* Other components that need access to the map */}
       {children}
     </MapboxProvider>
